Converted some tests from legacy to the new test suite.  I've also added a new mark for tests that use 'pipenv run'
